+++ src/providers/hiro.rs
@@ -12,7 +12,7 @@ use {
     },
     hyper::{client::HttpConnector, http, Client, Method},
     hyper_tls::HttpsConnector,
-    serde::Serialize,
+    serde::{Deserialize, Serialize},
     std::collections::HashMap,
     tracing::debug,
 };
@@ -23,9 +23,23 @@ pub struct HiroProvider {
     pub supported_chains: HashMap<String, String>,
 }
 
+#[derive(Debug, Serialize, Deserialize)]
+#[serde(rename_all = "snake_case")]
+pub enum SupportedMethods {
+    StacksTransactions,
+    StacksAccounts,
+    StacksExtendedNonces,
+    HiroFeesTransaction,
+}
+
 #[derive(Debug, Serialize)]
 pub struct TransactionsRequest {
     pub tx: String,
+}
+
+#[derive(Debug, Serialize)]
+pub struct FeesTransactionRequest {
+    pub transaction_payload: String,
 }
 
 impl Provider for HiroProvider {
@@ -49,44 +63,17 @@ impl RateLimited for HiroProvider {
     }
 }
 
-#[async_trait]
-impl RpcProvider for HiroProvider {
-    /// Proxies the request to the Stacks `/v2/transactions` endpoint only
-    /// using the JSON-RPC schema with the `stacks_transactions` method
-    /// and a single parameter as `tx`.
-    #[tracing::instrument(skip(self, body), fields(provider = %self.provider_kind()), level = "debug")]
-    async fn proxy(&self, chain_id: &str, body: hyper::body::Bytes) -> RpcResult<Response> {
+impl HiroProvider {
+    // Send request to the Stacks `/v2/transactions` endpoint
+    async fn transactions(&self, chain_id: String, tx: String) -> RpcResult<Response> {
         let uri = self
             .supported_chains
-            .get(chain_id)
+            .get(&chain_id)
             .ok_or(RpcError::ChainNotFound)?;
         let uri = format!("{}/v2/transactions", uri.trim_end_matches('/'));
         let uri = uri.parse::<hyper::Uri>().map_err(|_| {
             RpcError::InvalidParameter("Failed to parse URI for stacks_transactions".into())
         })?;
-
-        let json_rpc_request: JsonRpcRequest = serde_json::from_slice(&body)
-            .map_err(|_| RpcError::InvalidParameter("Invalid JSON-RPC schema provided".into()))?;
-
-        if json_rpc_request.method != "stacks_transactions".into() {
-            return Err(RpcError::InvalidParameter(
-                "Invalid method provided. Only 'stacks_transactions' is currently supported".into(),
-            ));
-        }
-
-        // Create the request body for stacks transactions endpoint schema
-        // by extracting the first parameter from the JSON-RPC request and using it as `tx`.
-        let tx_param = json_rpc_request
-            .params
-            .as_array()
-            .and_then(|arr| arr.first())
-            .unwrap_or(&json_rpc_request.params);
-
-        let tx = if let serde_json::Value::String(s) = tx_param {
-            s.clone()
-        } else {
-            tx_param.to_string()
-        };
 
         let stacks_transactions_request = serde_json::to_string(&TransactionsRequest { tx })
             .map_err(|e| {
@@ -107,7 +94,7 @@ impl RpcProvider for HiroProvider {
             if response.error.is_some() && status.is_success() {
                 debug!(
                     "Strange: provider returned JSON RPC error, but status {status} is success: \
-                 Hiro: {response:?}"
+                 Hiro transactions: {response:?}"
                 );
             }
         }
@@ -117,6 +104,222 @@ impl RpcProvider for HiroProvider {
             .headers_mut()
             .insert("Content-Type", HeaderValue::from_static("application/json"));
         Ok(response)
+    }
+
+    // Send request to the Stacks `/v2/accounts` endpoint
+    async fn accounts(&self, chain_id: String, principal: String) -> RpcResult<Response> {
+        let uri = self
+            .supported_chains
+            .get(&chain_id)
+            .ok_or(RpcError::ChainNotFound)?;
+        let uri = format!("{}/v2/accounts/{}", uri.trim_end_matches('/'), principal);
+        let uri = uri.parse::<hyper::Uri>().map_err(|_| {
+            RpcError::InvalidParameter("Failed to parse URI for stacks_accounts".into())
+        })?;
+
+        let hyper_request = hyper::http::Request::builder()
+            .method(Method::GET)
+            .uri(uri)
+            .header("Content-Type", "application/json")
+            .body(hyper::body::Body::empty())?;
+
+        let response = self.client.request(hyper_request).await?;
+        let status = response.status();
+        let body = hyper::body::to_bytes(response.into_body()).await?;
+
+        if let Ok(response) = serde_json::from_slice::<jsonrpc::Response>(&body) {
+            if response.error.is_some() && status.is_success() {
+                debug!(
+                    "Strange: provider returned JSON RPC error, but status {status} is success: \
+                 Hiro accounts: {response:?}"
+                );
+            }
+        }
+
+        let mut response = (status, body).into_response();
+        response
+            .headers_mut()
+            .insert("Content-Type", HeaderValue::from_static("application/json"));
+        Ok(response)
+    }
+
+    // Send request to the Hiro `/v2/fees/transaction` endpoint
+    async fn fees_transaction(
+        &self,
+        chain_id: String,
+        transaction_payload: String,
+    ) -> RpcResult<Response> {
+        let uri = self
+            .supported_chains
+            .get(&chain_id)
+            .ok_or(RpcError::ChainNotFound)?;
+        let uri = format!("{}/v2/fees/transaction", uri.trim_end_matches('/'));
+        let uri = uri.parse::<hyper::Uri>().map_err(|_| {
+            RpcError::InvalidParameter("Failed to parse URI for hiro_fees_transaction".into())
+        })?;
+
+        let hiro_fees_transaction_request = serde_json::to_string(&FeesTransactionRequest {
+            transaction_payload,
+        })
+        .map_err(|e| {
+            RpcError::InvalidParameter(format!("Failed to serialize fees transaction: {}", e))
+        })?;
+
+        let hyper_request = hyper::http::Request::builder()
+            .method(Method::POST)
+            .uri(uri)
+            .header("Content-Type", "application/json")
+            .body(hyper::body::Body::from(hiro_fees_transaction_request))?;
+
+        let response = self.client.request(hyper_request).await?;
+        let status = response.status();
+        let body = hyper::body::to_bytes(response.into_body()).await?;
+
+        if let Ok(response) = serde_json::from_slice::<jsonrpc::Response>(&body) {
+            if response.error.is_some() && status.is_success() {
+                debug!(
+                    "Strange: provider returned JSON RPC error, but status {status} is success: \
+                 Hiro fees transaction: {response:?}"
+                );
+            }
+        }
+
+        let mut response = (status, body).into_response();
+        response
+            .headers_mut()
+            .insert("Content-Type", HeaderValue::from_static("application/json"));
+        Ok(response)
+    }
+
+    // Send request to the Stacks `/extended/v1/address/<principal>/nonces` endpoint
+    async fn extended_nonces(&self, chain_id: String, principal: String) -> RpcResult<Response> {
+        let uri = self
+            .supported_chains
+            .get(&chain_id)
+            .ok_or(RpcError::ChainNotFound)?;
+        let uri = format!(
+            "{}/extended/v1/address/{}/nonces",
+            uri.trim_end_matches('/'),
+            principal
+        );
+        let uri = uri.parse::<hyper::Uri>().map_err(|_| {
+            RpcError::InvalidParameter("Failed to parse URI for stacks_extended_nonces".into())
+        })?;
+
+        let hyper_request = hyper::http::Request::builder()
+            .method(Method::GET)
+            .uri(uri)
+            .header("Content-Type", "application/json")
+            .body(hyper::body::Body::empty())?;
+
+        let response = self.client.request(hyper_request).await?;
+        let status = response.status();
+        let body = hyper::body::to_bytes(response.into_body()).await?;
+
+        if let Ok(response) = serde_json::from_slice::<jsonrpc::Response>(&body) {
+            if response.error.is_some() && status.is_success() {
+                debug!(
+                    "Strange: provider returned JSON RPC error, but status {status} is success: \
+                 Stacks extended nonces: {response:?}"
+                );
+            }
+        }
+
+        let mut response = (status, body).into_response();
+        response
+            .headers_mut()
+            .insert("Content-Type", HeaderValue::from_static("application/json"));
+        Ok(response)
+    }
+}
+
+#[async_trait]
+impl RpcProvider for HiroProvider {
+    /// Proxies the request to the Stacks endpoints
+    /// using the JSON-RPC schema `method` to map the endpoint and parameters.
+    #[tracing::instrument(skip(self, body), fields(provider = %self.provider_kind()), level = "debug")]
+    async fn proxy(&self, chain_id: &str, body: hyper::body::Bytes) -> RpcResult<Response> {
+        let json_rpc_request: JsonRpcRequest = serde_json::from_slice(&body)
+            .map_err(|_| RpcError::InvalidParameter("Invalid JSON-RPC schema provided".into()))?;
+
+        let method: SupportedMethods = serde_json::from_value(serde_json::Value::String(
+            (*json_rpc_request.method).to_string(),
+        ))
+        .map_err(|e| RpcError::InvalidParameter(format!("Invalid method provided: {:?}", e)))?;
+
+        match method {
+            SupportedMethods::StacksTransactions => {
+                // Create the request body for stacks transactions endpoint schema
+                // by extracting the first parameter from the JSON-RPC request and using it as `tx`.
+                let tx_param = json_rpc_request
+                    .params
+                    .as_array()
+                    .and_then(|arr| arr.first())
+                    .unwrap_or(&json_rpc_request.params);
+
+                let tx = if let serde_json::Value::String(s) = tx_param {
+                    s.clone()
+                } else {
+                    tx_param.to_string()
+                };
+
+                return self.transactions(chain_id.to_string(), tx).await;
+            }
+            SupportedMethods::StacksAccounts => {
+                // Create the request body for stacks accounts endpoint schema
+                // by extracting the first parameter from the JSON-RPC request and using it as `principal`.
+                let tx_param = json_rpc_request
+                    .params
+                    .as_array()
+                    .and_then(|arr| arr.first())
+                    .unwrap_or(&json_rpc_request.params);
+
+                let tx = if let serde_json::Value::String(s) = tx_param {
+                    s.clone()
+                } else {
+                    tx_param.to_string()
+                };
+
+                return self.accounts(chain_id.to_string(), tx).await;
+            }
+            SupportedMethods::HiroFeesTransaction => {
+                // Create the request body for hiro fees transactions endpoint schema
+                // by extracting the first parameter from the JSON-RPC request
+                // and using it as `transaction_payload`.
+                let tx_param = json_rpc_request
+                    .params
+                    .as_array()
+                    .and_then(|arr| arr.first())
+                    .unwrap_or(&json_rpc_request.params);
+
+                let transaction_payload = if let serde_json::Value::String(s) = tx_param {
+                    s.clone()
+                } else {
+                    tx_param.to_string()
+                };
+
+                return self
+                    .fees_transaction(chain_id.to_string(), transaction_payload)
+                    .await;
+            }
+            SupportedMethods::StacksExtendedNonces => {
+                // Create the request body for stacks extended nonces endpoint schema
+                // by extracting the first parameter from the JSON-RPC request and using it as `principal`.
+                let tx_param = json_rpc_request
+                    .params
+                    .as_array()
+                    .and_then(|arr| arr.first())
+                    .unwrap_or(&json_rpc_request.params);
+
+                let tx = if let serde_json::Value::String(s) = tx_param {
+                    s.clone()
+                } else {
+                    tx_param.to_string()
+                };
+
+                return self.extended_nonces(chain_id.to_string(), tx).await;
+            }
+        }
     }
 }