+++ tests/functional/mod.rs
@@ -1,4 +1,5 @@
 mod bundler;
 mod database;
 mod http;
+mod sessions;
 mod websocket;